third_party/blink/web_tests/fast/dom/shadow/event-path-for-user-agent-shadow-tree-expected.txt


Dispaching a click event on #details-child

event.composedPath() on node #details-child
#details-child, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 7

event.composedPath() on node #details
#details-child, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 7

event.composedPath() on node #sandbox
#details-child, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 7

Dispaching a click event on #summary-child

event.composedPath() on node #summary-child
#summary-child, #summary, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

event.composedPath() on node #summary
#summary-child, #summary, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

event.composedPath() on node #details
#summary-child, #summary, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8

event.composedPath() on node #sandbox
#summary-child, #summary, #details, #sandbox, [object HTMLBodyElement], [object HTMLHtmlElement], [object HTMLDocument], [object Window], length: 8
PASS successfullyParsed is true

TEST COMPLETE
